--- a/Gemfile
+++ b/Gemfile
@@ -147,3 +147,5 @@ gem 'rest-client'
 gem 'rtesseract'
 
 gem 'multipart-post'
+
+gem 'redis-rails'