--- a/config/environments/digcollhyrax02.rb
+++ b/config/environments/digcollhyrax02.rb
@@ -53,7 +53,11 @@ Rails.application.configure do
   #
   #  config.cache_store = :null_store
   #end
+
+  # AR: Use redis for caching
   config.action_controller.perform_caching = true
+  config.cache_store = :redis_store, "redis://localhost:6379/0/cache"
+  config.middleware.insert_before ActionDispatch::Static, NoCacheMiddleware, [/concern\/.*\/.*\/edit/, /concern\/.*\/.*\/new/, /dashboard\/.*\/.*\/edit/, /uploaded_collection_thumbnails/]
 
   # Don't care if the mailer can't send.
   config.action_mailer.raise_delivery_errors = false
@@ -105,5 +109,4 @@ Rails.application.configure do
 
   # JL: For Devise account locking
   Rails.application.routes.default_url_options[:host] = 'http://digcoll-hyrax02.tcd.ie'
-
 end