--- a/config/features.rb
+++ b/config/features.rb
@@ -1,0 +1,2 @@
+cache_work_iiif_manifest:
+  enabled: true