This is currently live on the secondary server, and the changes will persist once deployed.

When this PR is deployed to the primary server, it will configure the application to use Redis for caching via:  
`cache_store = :redis_store` as well as enable the cache_work_iiif_manifest for manifest caching.

# Story
Since Rails 5.1.6.2 does not include out-of-the-box support for `redis_cache_store`, we use the `redis-rails` gem instead. The cache store configuration for both the primary and secondary environments has been updated to:

```ruby
config.cache_store = :redis_store, "redis://localhost:6379/0/cache"
```

Note that this configuration does not rely on environment variables, as initially assumed.

# IIIF Manifest Caching Performance Gains

As a follow-up, we enabled the `cache_work_iiif_manifest` feature via Flipflop and confirmed it is active in production. This change leverages Redis-backed Rails caching to store IIIF manifest JSON responses.

## Benchmark Summary
We ran 20 total manifest load tests (10 before caching, 10 after). Here's the comparison:

| Metric                    | Before Caching | After Caching | Improvement |
|--------------------------|----------------|----------------|-------------|
| Avg manifest load time   | ~1.4 seconds    | ~0.56 seconds  | ~60% faster |
| Fastest manifest load    | ~0.53 seconds   | ~0.528 seconds | ~0% (already fast) |
| Slowest manifest load    | ~4.15 seconds   | ~0.58 seconds  | ~86% faster |

*Note:* First loads are still slower (~4.15s pre-caching), but caching improves performance dramatically for all subsequent requests.

This enhancement provides a much faster IIIF viewer experience and reduces repeated backend load.